nixos-enter: use the correct bash executable
Consider for instance a case in which only a single `--system` parameter
is passed to `nixos-enter`. Then during parsing, the `$system` variable
will be set correctly, but `$command` will not be changed from the
initial value set before parsing. Therefore, the value of `--system`
will not be used while picking the `bash` to be used.